wt88047: Fix dark video capture
There is an issue with some of the libchromatix libraries (specifically the ov2680 and the ov8865_q8v18a) that causes video capture to be extremely dark under good light conditions. This commit fixes it by replacing those faulty libraries with the ones that work as expected (ov2680_5987fhq and ov8865_sunny).